Example weights for the working and trained 'Cart' NN
These are the saved weights of the NN trained in the ipynb file. This NN can play the game for 312 actions in a row before it loses, and maybe even longer.
Feel free to use these weights to check it out yourself and maybe keep training the network and improve its performance.